+++ News_springbootbackend/src/main/java/com/example/news_springbootbackend/entity/SecurityUser.java
@@ -18,7 +18,7 @@ public class SecurityUser implements UserDetails {
         this.user = user;
     }
 
-    public Long getID() {
+    public int getID() {
         return user.getId();
     }
